+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to have no row headers if list has only one semantic column.
+
+## 1.14.0 - (March 29, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.13.0 - (March 25, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.12.0 - (March 21, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.11.0 - (March 19, 2024)
 
 * Changed